# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API server for the REGS Online Judge system.",
        "title": "REGS Online Judge API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/api",
    "paths": {
...
        "/submissions": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Retrieves a list of all submissions made by the currently authenticated user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Submissions"
                ],
                "summary": "Get personal submission history",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Submission"
                            }
                        }
                    },
                    "401": {
                        "description": "未授權的操作",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "無法取得提交紀錄",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Upload a .zip file containing source code for a specific problem.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Submissions"
                ],
                "summary": "Submit code for judging",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Problem ID (e.g., p1001)",
                        "name": "problem_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Source code as a .zip file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "提交成功，開始評測",
                        "schema": {
                            "type": "object",
                            "properties": {
                                " operatorId": {
                                    "type": "string"
                                },
                                " userId": {
                                    "type": "integer"
                                },
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "請求錯誤",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "伺服器內部錯誤",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/submissions/{operatorId}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Retrieves the current status and result of a specific submission.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Submissions"
                ],
                "summary": "Get submission status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Operator ID of the submission",
                        "name": "operatorId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                " created_at": {
                                    "type": "string"
                                },
                                " run_memory": {
                                    "type": "integer"
                                },
                                " run_time": {
                                    "type": "integer"
                                },
                                " status": {
                                    "type": "string"
                                },
                                "operatorId": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "找不到該筆評測紀錄",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/submissions/{operatorId}/source": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Downloads the original .zip file for a submission. Only accessible by the owner or an admin.",
                "produces": [
                    "application/zip"
                ],
                "tags": [
                    "Submissions"
                ],
                "summary": "Download submission source code",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Operator ID of the submission",
                        "name": "operatorId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The submission's source code as a .zip file",
                        "schema": {
                            "type": "file"
                        }
                    }
                }
            }
        },
...
        "/users/{user_id}/submissions": {
            "get": {
                "description": "Retrieves a list of all submissions made by a specific user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Submissions"
                ],
                "summary": "Get a specific user's submission history",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Submission"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.Problem": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_visible": {
                    "type": "boolean"
                },
                "memory_limit": {
                    "type": "integer"
                },
                "testcase_path": {
                    "type": "string"
                },
                "time_limit": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "models.Submission": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "operatorID": {
                    "type": "string"
                },
                "problem": {
                    "$ref": "#/definitions/models.Problem"
                },
                "problemID": {
                    "type": "string"
                },
                "run_memory": {
                    "type": "integer"
                },
                "run_time": {
                    "type": "integer"
                },
                "score": {
                    "description": "實際得分",
                    "type": "integer"
                },
                "status": {
                    "description": "Pending, AC, WA, CE, SE, RE, TLE",
                    "type": "string"
                },
                "total_score": {
                    "description": "該題滿分",
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/models.User"
                },
                "userID": {
                    "type": "integer"
                }
            }
        },
        "models.User": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "role": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}